// Boardlight/src/api/generated/models/schoolIndexGradesResponse.ts
/**
 * Generated by orval v6.17.0 🍺
 * Do not edit manually.
 * Blueboard
 * OpenAPI spec version: 4.1.0
 */
import type { SchoolIndexGradesResponseGrade } from './schoolIndexGradesResponseGrade';

export interface SchoolIndexGradesResponse {
  subject?: string | null;
  grades?: SchoolIndexGradesResponseGrade[] | null;
}