// Boardlight/src/api/generated/models/shopIndexLolosResponse.ts
/**
 * Generated by orval v6.16.0 🍺
 * Do not edit manually.
 * Blueboard
 * OpenAPI spec version: v4.0.0
 */

export interface ShopIndexLolosResponse {
  id?: number;
  userId?: string;
  isSpent?: boolean;
  loloType?: string | null;
  reason?: string | null;
  createdAt?: string;
  updatedAt?: string;
}