// Boardlight/src/api/generated/models/shopIndexOwnLolosResponse.ts
/**
 * Generated by orval v6.17.0 🍺
 * Do not edit manually.
 * Blueboard
 * OpenAPI spec version: 4.1.0
 */
import type { ShopIndexOwnLolosResponseCoin } from './shopIndexOwnLolosResponseCoin';

export interface ShopIndexOwnLolosResponse {
  balance?: number;
  coins?: ShopIndexOwnLolosResponseCoin[] | null;
}